move mock_import_error to fixture
